rename and reorg DB handler
