+++ cellpack/bin/upload.py
@@ -1,7 +1,7 @@
 from enum import Enum
 import fire
 from cellpack.autopack.FirebaseHandler import FirebaseHandler
-from cellpack.autopack.DBRecipeHandler import DBHandler
+from cellpack.autopack.DBRecipeHandler import DBUploader
 
 from cellpack.autopack.loaders.recipe_loader import RecipeLoader
 
@@ -26,7 +26,7 @@ def upload(
         recipe_loader = RecipeLoader(recipe_path)
         recipe_full_data = recipe_loader.recipe_data
         recipe_meta_data = recipe_loader.get_only_recipe_metadata()
-        recipe_db_handler = DBHandler(db_handler)
+        recipe_db_handler = DBUploader(db_handler)
         recipe_db_handler.upload_recipe(recipe_meta_data, recipe_full_data)
